Add new tags attribute to project
This change adds "tags" as a new attribute for projects by
allowing projects to be "tagged" with a list of simple strings.
This will allow an easy method of searching and categorizing
projects by simple name values.

This change consists of adding new API links to modify the tags
of a project, adding tags as a property of project in the SQL
object, and updating any instances of project/domains to also
contain a list of tags.

Change-Id: I0983a5b592aaffa9cef38c74c038fa60a176786c
Partially-Implements: bp project-tags
Co-Authored-By: Jaewoo Park <jp655p@att.com>
Co-Authored-By: Nicolas Helgeson <nh202b@att.com>
Depends-On: I00f094a5584be40ab477cbf680a5f6d1afb4d21b